test: move all filter tests together
