+++ src/test/scala/org/bblfsh/client/v2/BblfshClientBaseTest.scala
@@ -1,12 +1,12 @@
 package org.bblfsh.client.v2
 
 import gopkg.in.bblfsh.sdk.v2.protocol.driver.ParseResponse
-import org.scalatest.{BeforeAndAfter, BeforeAndAfterAll, FlatSpec, Matchers}
+import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach, FlatSpec, Matchers}
 
 import scala.io.Source
 
 class BblfshClientBaseTest extends FlatSpec
-  with BeforeAndAfter
+  with BeforeAndAfterEach
   with BeforeAndAfterAll
   with Matchers {
 
@@ -14,7 +14,7 @@ class BblfshClientBaseTest extends FlatSpec
   val fileName = "src/test/resources/SampleJavaFile.java"
   var resp: ParseResponse = _
 
-  before {
+  override def beforeEach() = {
     val fileContent = Source.fromFile(fileName).getLines.mkString("\n")
     resp = client.parse(fileName, fileContent)
   }
