+++ src/test/scala/org/bblfsh/client/v2/BblfshClientParseTest.scala
@@ -2,8 +2,6 @@ package org.bblfsh.client.v2
 
 import java.nio.ByteBuffer
 
-import gopkg.in.bblfsh.sdk.v2.protocol.driver.Mode
-
 import scala.io.Source
 
 class BblfshClientParseTest extends BblfshClientBaseTest {
@@ -13,42 +11,6 @@ class BblfshClientParseTest extends BblfshClientBaseTest {
   "Parsed UAST for .java file" should "not be empty" in {
     assert(resp != null)
     assert(resp.errors.isEmpty)
-  }
-
-  "Filtering UAST" should "work in Native mode" in {
-    val fileContent = Source.fromFile(fileName).getLines.mkString("\n")
-    val resp = client.parse(fileName, fileContent, Mode.NATIVE)
-    val node = resp.get
-
-    val iter = BblfshClient.filter(node, "//SimpleName")
-    iter.toList should have size (10) // number of Identifiers in the file
-    iter.close()
-  }
-
-// TODO(#110) implement value type returns
-//  "Filtering UAST" should "work for Value types" in {
-//    val iter = BblfshClient.filterNumber(resp.get, "count(//*)")
-//    iter.toList should have size (517) // total number of nodes (not the number of results which is 1)
-//  }
-
-  "Filtering UAST" should "work in Annotated mode" in {
-    val fileContent = Source.fromFile(fileName).getLines.mkString("\n")
-    val resp = client.parse(fileName, fileContent, Mode.ANNOTATED)
-    val node = resp.get
-
-    val iter = BblfshClient.filter(node, "//SimpleName[@role='Call']")
-    iter.toList should have size (1) // number of function called in the file
-    iter.close()
-  }
-
-  "Filtering UAST" should "work in Semantic mode" in {
-    val fileContent = Source.fromFile(fileName).getLines.mkString("\n")
-    val resp = client.parse(fileName, fileContent, Mode.SEMANTIC)
-    val node = resp.get
-
-    val iter = BblfshClient.filter(node, "//uast:Identifier[@role='Call']")
-    iter.toList should have size (1) // number of function called in the file
-    iter.close()
   }
 
   "Decoded UAST after parsing" should "not be NULL" in {
