+++ src/test/scala/org/bblfsh/client/v2/FilterManagedTest.scala
@@ -1,12 +1,13 @@
 package org.bblfsh.client.v2
 
+import gopkg.in.bblfsh.sdk.v2.protocol.driver.Mode
 import org.bblfsh.client.v2.libuast.Libuast
-import org.scalatest.{BeforeAndAfter, BeforeAndAfterAll, FlatSpec, Matchers}
 
-class FilterManagedTest extends FlatSpec
-  with Matchers
-  with BeforeAndAfter
-  with BeforeAndAfterAll {
+import scala.io.Source
+
+class FilterManagedTest extends BblfshClientBaseTest {
+
+  import BblfshClient._ // enables uast.* methods
 
   var ctx: Context = _
   val managedRoot = JArray(
@@ -19,15 +20,18 @@ class FilterManagedTest extends FlatSpec
     ))
 
   override def beforeAll() = {
+    super.beforeAll()
     System.err.println(s"Libuast.loaded: ${Libuast.loaded}")
     // to load native JNI lib \wo the full client
   }
 
-  before {
+  override def beforeEach() = {
+    super.beforeEach()
     ctx = Context()
   }
 
   override def afterAll() = {
+    super.afterAll()
     System.runFinalization()
     System.gc()
   }
@@ -41,6 +45,42 @@ class FilterManagedTest extends FlatSpec
 
     it.close()
     it.hasNext() should be(false)
+  }
+
+  "Filtering UAST" should "work in Native mode" in {
+    val fileContent = Source.fromFile(fileName).getLines.mkString("\n")
+    val resp = client.parse(fileName, fileContent, Mode.NATIVE)
+    val node = resp.get
+
+    val iter = BblfshClient.filter(node, "//SimpleName")
+    iter.toList should have size (10) // number of Identifiers in the file
+    iter.close()
+  }
+
+  // TODO(#110) implement value type returns
+  //  "Filtering UAST" should "work for Value types" in {
+  //    val iter = BblfshClient.filterNumber(resp.get, "count(//*)")
+  //    iter.toList should have size (517) // total number of nodes (not the number of results which is 1)
+  //  }
+
+  "Filtering UAST" should "work in Annotated mode" in {
+    val fileContent = Source.fromFile(fileName).getLines.mkString("\n")
+    val resp = client.parse(fileName, fileContent, Mode.ANNOTATED)
+    val node = resp.get
+
+    val iter = BblfshClient.filter(node, "//SimpleName[@role='Call']")
+    iter.toList should have size (1) // number of function called in the file
+    iter.close()
+  }
+
+  "Filtering UAST" should "work in Semantic mode" in {
+    val fileContent = Source.fromFile(fileName).getLines.mkString("\n")
+    val resp = client.parse(fileName, fileContent, Mode.SEMANTIC)
+    val node = resp.get
+
+    val iter = BblfshClient.filter(node, "//uast:Identifier[@role='Call']")
+    iter.toList should have size (1) // number of function called in the file
+    iter.close()
   }
 
 }
